added unit test for glossary term page
