+++ openmetadata-ui/src/main/resources/ui/src/components/GlossaryTerms/GlossaryTerms.test.tsx
@@ -11,11 +11,12 @@
  *  limitations under the License.
  */
 
-import { findByText, getByTestId, render } from '@testing-library/react';
+import { act, fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import {
   mockedAssetData,
   mockedGlossaryTerms,
+  MOCK_ASSETS_DATA,
 } from '../../mocks/Glossary.mock';
 import { OperationPermission } from '../PermissionProvider/PermissionProvider.interface';
 import GlossaryTerms from './GlossaryTermsV1.component';
@@ -68,6 +69,12 @@ jest.mock('react-router-dom', () => ({
   }),
 }));
 
+jest.mock('rest/miscAPI', () => ({
+  searchData: jest
+    .fn()
+    .mockImplementation(() => Promise.resolve(MOCK_ASSETS_DATA)),
+}));
+
 jest.mock('components/Tag/TagsContainer/tags-container', () => {
   return jest.fn().mockReturnValue(<>Tags-container component</>);
 });
@@ -80,48 +87,6 @@ jest.mock('../common/rich-text-editor/RichTextEditorPreviewer', () => {
   return jest.fn().mockReturnValue(<p>RichTextEditorPreviewer</p>);
 });
 
-jest.mock('antd', () => ({
-  ...jest.requireActual('antd'),
-  Card: jest
-    .fn()
-    .mockImplementation(({ children, ...props }) => (
-      <div {...props}>{children}</div>
-    )),
-  Col: jest
-    .fn()
-    .mockImplementation(({ children, ...props }) => (
-      <div {...props}>{children}</div>
-    )),
-  Row: jest
-    .fn()
-    .mockImplementation(({ children, ...props }) => (
-      <div {...props}>{children}</div>
-    )),
-  Divider: jest
-    .fn()
-    .mockImplementation(({ children, ...props }) => (
-      <div {...props}>{children}</div>
-    )),
-  Typography: {
-    Text: jest.fn().mockImplementation(({ children }) => <div>{children}</div>),
-  },
-  Space: jest
-    .fn()
-    .mockImplementation(({ children, ...props }) => (
-      <div {...props}>{children}</div>
-    )),
-  Input: jest
-    .fn()
-    .mockImplementation(({ children, ...props }) => (
-      <div {...props}>{children}</div>
-    )),
-  Button: jest
-    .fn()
-    .mockImplementation(({ children, ...props }) => (
-      <div {...props}>{children}</div>
-    )),
-}));
-
 jest.mock('./SummaryDetail', () =>
   jest.fn().mockReturnValue(<div>SummaryDetails</div>)
 );
@@ -133,6 +98,12 @@ jest.mock('./tabs/GlossaryTermSynonyms', () =>
 );
 jest.mock('./tabs/GlossaryTermReferences', () =>
   jest.fn().mockReturnValue(<div>GlossaryTermReferencesComponent</div>)
+);
+jest.mock('./tabs/AssetsTabs.component', () =>
+  jest.fn().mockReturnValue(<div>AssetsTabs</div>)
+);
+jest.mock('components/Glossary/GlossaryTermTab/GlossaryTermTab.component', () =>
+  jest.fn().mockReturnValue(<div>GlossaryTermTab</div>)
 );
 
 const mockProps = {
@@ -154,30 +125,60 @@ const mockProps = {
 };
 
 describe('Test Glossary-term component', () => {
-  it('Should render Glossary-term component', () => {
-    const { container } = render(<GlossaryTerms {...mockProps} />);
+  it('Should render Glossary-term component', async () => {
+    await act(async () => {
+      render(<GlossaryTerms {...mockProps} />);
+    });
 
-    const glossaryTerm = getByTestId(container, 'glossary-term');
-
-    expect(glossaryTerm).toBeInTheDocument();
-  });
-
-  it('Should render Tags-container', async () => {
-    const { container } = render(<GlossaryTerms {...mockProps} />);
-
-    const tagsContainer = await findByText(
-      container,
-      /Tags-container component/i
-    );
-
-    expect(tagsContainer).toBeInTheDocument();
-  });
-
-  it('Should render Description', async () => {
-    const { container } = render(<GlossaryTerms {...mockProps} />);
-
-    const description = await findByText(container, /Description component/i);
+    const glossaryTerm = screen.getByTestId('glossary-term');
+    const tagsContainer = await screen.findByText(/Tags-container component/i);
+    const description = await screen.findByText(/Description component/i);
+    const tabs = await screen.findAllByRole('tab');
 
     expect(description).toBeInTheDocument();
+    expect(
+      await screen.findByText('RelatedTermsComponent')
+    ).toBeInTheDocument();
+    expect(
+      await screen.findByText('GlossaryTermSynonymsComponent')
+    ).toBeInTheDocument();
+    expect(
+      await screen.findByText('GlossaryTermReferencesComponent')
+    ).toBeInTheDocument();
+    expect(await screen.findByText('Mocked User')).toBeInTheDocument();
+    expect(tagsContainer).toBeInTheDocument();
+    expect(glossaryTerm).toBeInTheDocument();
+    expect(tabs).toHaveLength(3);
+    expect(tabs.map((tab) => tab.textContent)).toStrictEqual([
+      'label.summary',
+      'label.asset-plural1', // 1 added as its count for assets
+      'label.glossary-term-plural',
+    ]);
+  });
+
+  it('onClick of assets tab, it should render properly', async () => {
+    await act(async () => {
+      render(<GlossaryTerms {...mockProps} />);
+    });
+    const tabs = await screen.findAllByRole('tab');
+    await act(async () => {
+      fireEvent.click(tabs[1]);
+    });
+
+    expect(tabs[1].textContent).toStrictEqual('label.asset-plural1');
+    expect(await screen.findByText('AssetsTabs')).toBeInTheDocument();
+  });
+
+  it('onClick of glossary term tab, it should render properly', async () => {
+    await act(async () => {
+      render(<GlossaryTerms {...mockProps} />);
+    });
+    const tabs = await screen.findAllByRole('tab');
+    await act(async () => {
+      fireEvent.click(tabs[2]);
+    });
+
+    expect(tabs[2].textContent).toStrictEqual('label.glossary-term-plural');
+    expect(await screen.findByText('GlossaryTermTab')).toBeInTheDocument();
   });
 });
